refactor: rename page routes and create initial card route
