+++ src/app/(private)/conta/page.tsx
@@ -4,7 +4,7 @@ import UpdatePasswordForm from "./components/update-password-form";
 import UpdateUserInfosForm from "./components/update-user-infos-form";
 import UserDataSection from "./components/user-data-section";
 
-export default function MyAccountScreen() {
+export default function AccountScreen() {
   return (
     <AuthGuard requireAuth>
       <Column className="space-y-4 items-center w-full justify-center h-fit sm:max-w-5xl mx-auto mt-16 lg:py-4">
