+++ src/app/(private)/transacoes/page.tsx
@@ -1,7 +1,7 @@
 import AuthGuard from "@/src/components/utils/auth-guard";
 import FinancialDashboard from "./components/financial-dashboard";
 
-export default function Home() {
+export default function TransactionsScreen() {
   return (
     <AuthGuard requireAuth>
       <FinancialDashboard />
