+++ src/app/global-not-found.tsx
@@ -27,7 +27,7 @@ export default function GlobalNotFound() {
             A página que você está tentando acessar não existe
           </p>
         </Column>
-        <Link href="/dashboard">
+        <Link href="/transacoes">
           <Button className="cursor-pointer" type="button">
             Voltar para o Finance Tracker
           </Button>
