+++ src/components/header.tsx
@@ -13,7 +13,7 @@ const Header = () => {
   return (
     <header className="w-full bg-black h-16 text-white p-4 fixed top-0 left-0 z-10">
       <Row className="max-w-5xl mx-auto justify-between items-center">
-        <Link href="/dashboard">
+        <Link href="/transacoes">
           <h1 className="text-lg font-bold !text-green-600">Finance Tracker</h1>
         </Link>
         <UserMenu />
