+++ src/components/user-menu.tsx
@@ -2,6 +2,7 @@
 
 import { useMutation } from "@tanstack/react-query";
 import {
+  CreditCard,
   Loader2Icon,
   LogOut,
   Settings,
@@ -66,13 +67,19 @@ const UserMenu = () => {
         </DropdownMenuLabel>
         <DropdownMenuSeparator />
         <DropdownMenuGroup>
-          <DropdownMenuItem onClick={() => route.push("/dashboard")}>
+          <DropdownMenuItem onClick={() => route.push("/transacoes")}>
             Minhas Transações
             <DropdownMenuShortcut className="space-x-1">
               <Wallet />
             </DropdownMenuShortcut>
           </DropdownMenuItem>
-          <DropdownMenuItem onClick={() => route.push("/minha-conta")}>
+          <DropdownMenuItem onClick={() => route.push("/cartoes")}>
+            Meus Cartões
+            <DropdownMenuShortcut className="space-x-1">
+              <CreditCard />
+            </DropdownMenuShortcut>
+          </DropdownMenuItem>
+          <DropdownMenuItem onClick={() => route.push("/conta")}>
             Minha Conta
             <DropdownMenuShortcut className="space-x-1">
               <Settings />
